refactoring: moving functions to separate file for better navigation
