+++ src/helpful_functions.cpp
@@ -1,4 +1,32 @@
-#include<iostream>
-
+#include <iostream>
+#include <vector>
+#include <array>
+#include <fstream>
+#include <string>
 
 uint16_t big_endian_to_small(uint16_t value) { return (((value & 0xff)<<8) | ((value & 0xff00)>>8)); }
+
+std::vector<std::pair<int, std::string>> load_configurations(const std::string& name)
+{
+    std::string text;
+    std::ifstream filename{name};
+    std::vector<std::pair<int, std::string>> pairs;
+    if(filename.is_open())
+    {
+        int position;
+        while(getline(filename, text))
+        {
+            position = text.find_first_of(' ');
+            pairs.push_back({
+                std::stoi(text.substr(0, position), nullptr, 16),
+                text.substr(position+1, text.size())});
+        }
+        filename.close();
+    }
+    else 
+    {
+        std::cout << "Unable to open file"; 
+        return std::vector<std::pair<int, std::string>>();
+    }
+    return pairs;
+}
